极端情况测试样例在script/extra_3.txt与extra_4.txt中，把他们拖到run.bat运行即可，我在SimulationSystem的90行加了个sleep帮助看清楚，不用可以注释一下。

+++ src/proj/Console.java
@@ -34,8 +34,8 @@ public class Console {
                 e.printStackTrace();
             }
         }
-        double range_x = input.nextInt();
-        double range_y = input.nextInt();
+        double range_x = input.nextDouble();
+        double range_y = input.nextDouble();
         this.range = new double[]{0, range_x, 0, range_y};
         int count = input.nextInt();
         cells = new Cell[count];
